Add files via upload
ChatClient connects via ChatServer to server. 
ChatServer sets up the server and waits to listen for clients.
ChatServerThread/ChatClientThread creates Threads for Server and Client
OneTimePad program used to process encryption and decryption of messages from clients
ChatClientGUI creates a Graphics User Interface of a ChatClient which also connects to the server and behaves like the client.

All messages go to server.